FEAT: Built and tested a basic lcore UDP function which will the foundation of our NSPK RTP framework after this commit. FIX: Fixed and refactored the build.sh script.
