+++ include/nspk_control.h
@@ -1,0 +1,3 @@
+#include <nspk.h>
+
+int lcore_main_control(void *arg);
