+++ include/nspk_rtp.h
@@ -1,1 +1,4 @@
 #pragma once
+
+int
+lcore_main_rtp(void *arg);
