+++ src/nspk_control.c
@@ -1,0 +1,28 @@
+#include <nspk_control.h>
+#include <tldk_utils/udp.h>
+
+int
+lcore_main_control(void *arg)
+{
+	int32_t rc;
+	uint32_t lcore;
+	struct lcore_prm *prm;
+
+	prm = arg;
+	lcore = rte_lcore_id();
+
+	RTE_LOG(NOTICE, USER1, "%s(lcore=%u) start\n",
+		__func__, lcore);
+
+	while (force_quit == 0) {
+		;
+	}
+
+	RTE_LOG(NOTICE, USER1, "%s(lcore=%u) finish\n",
+		__func__, lcore);
+
+	netfe_lcore_fini_udp();
+	netbe_lcore_clear();
+
+	return rc;
+}
